requirements.txt
python-nmap==0.7.1
requests==2.31.0
asyncio
reportlab==4.0.4
jinja2==3.1.2
flask==2.3.3
werkzeug==2.3.7
sqlite3
aiohttp==3.8.5
